+++ UdpHosts/GameServer/Entities/Carryable/CarryableEntity.cs
@@ -200,7 +200,6 @@ public class CarryableEntity : BaseAptitudeEntity, IAptitudeTarget
             AllowHostilePickupProp = AllowHostilePickup,
             PersonalFactionStanceProp = null,
             ScopeBubbleInfoProp = ScopeBubble
-
         };
     }
 }
