+++ UdpHosts/GameServer/Systems/Aptitude/Commands/Other/ApplyFreezeCommand.cs
@@ -48,5 +48,4 @@ public class ApplyFreezeCommand : ICommand
             }
         }
     }
-    
 }
