+++ UdpHosts/GameServer/Systems/Aptitude/Commands/Other/OrientationLockCommand.cs
@@ -48,5 +48,4 @@ public class OrientationLockCommand : ICommand
             }
         }
     }
-    
 }
